requirements.txt
cairocffi==1.7.1
CairoSVG==2.7.1
cffi==1.17.1
cssselect2==0.7.0
defusedxml==0.7.1
emoji==2.14.0
numpy==2.1.3
pillow==11.0.0
pycparser==2.22
pygame==2.6.1
pygame-cards==0.1.3
pygame_emojis==0.1.1
PyYAML==6.0.2
tinycss2==1.4.0
webencodings==0.5.1
